+++ app/controllers/devise/passwords_controller.rb
@@ -36,7 +36,7 @@ class Devise::PasswordsController < DeviseController
 
     if resource.errors.empty?
       resource.unlock_access! if unlockable?(resource)
-      if sign_in_after_reset_password?(resource)
+      if sign_in_after_reset_password?
         flash_message = resource.active_for_authentication? ? :updated : :updated_not_active
         set_flash_message!(:notice, flash_message)
         resource.after_database_authentication
@@ -58,7 +58,7 @@ class Devise::PasswordsController < DeviseController
     end
 
     def after_resetting_password_path_for(resource)
-      sign_in_after_reset_password?(resource) ? after_sign_in_path_for(resource) : new_session_path(resource_name)
+      sign_in_after_reset_password? ? after_sign_in_path_for(resource) : new_session_path(resource_name)
     end
 
     # The path used after sending reset password instructions
@@ -72,6 +72,12 @@ class Devise::PasswordsController < DeviseController
         set_flash_message(:alert, :no_token)
         redirect_to new_session_path(resource_name)
       end
+    end
+
+    # Check if the user should be signed in automatically after resetting the password.
+    def sign_in_after_reset_password?
+      setting = resource_class.sign_in_after_reset_password
+      setting.respond_to?(:call) ? setting.call(resource) : setting
     end
 
     # Check if proper Lockable module methods are present & unlock strategy