+++ test/integration/registerable_test.rb
@@ -98,10 +98,6 @@ class RegistrationTest < Devise::IntegrationTest
   end
 
   test 'a guest user cannot sign up with invalid information' do
-    # Dirty tracking behavior prevents email validations from being applied:
-    #    https://github.com/mongoid/mongoid/issues/756
-    (pending "Fails on Mongoid < 2.1"; break) if defined?(Mongoid) && Mongoid::VERSION.to_f < 2.1
-
     get new_user_registration_path
 
     fill_in 'email', with: 'invalid_email'
@@ -120,10 +116,6 @@ class RegistrationTest < Devise::IntegrationTest
   end
 
   test 'a guest should not sign up with email/password that already exists' do
-    # Dirty tracking behavior prevents email validations from being applied:
-    #    https://github.com/mongoid/mongoid/issues/756
-    (pending "Fails on Mongoid < 2.1"; break) if defined?(Mongoid) && Mongoid::VERSION.to_f < 2.1
-
     create_user
     get new_user_registration_path
 
@@ -181,6 +173,22 @@ class RegistrationTest < Devise::IntegrationTest
 
   test 'a signed in user should not be able to use the website after changing their password if config.sign_in_after_change_password is false' do
     swap Devise, sign_in_after_change_password: false do
+      sign_in_as_user
+      get edit_user_registration_path
+
+      fill_in 'password', with: '1234567890'
+      fill_in 'password confirmation', with: '1234567890'
+      fill_in 'current password', with: '12345678'
+      click_button 'Update'
+
+      assert_contain 'Your account has been updated successfully, but since your password was changed, you need to sign in again.'
+      assert_equal new_user_session_path, @request.path
+      assert_not warden.authenticated?(:user)
+    end
+  end
+
+  test 'a signed in user should not be able to use the website after changing their password if resource_class.sign_in_after_change_password is false' do
+    swap_model_config User, sign_in_after_change_password: false do
       sign_in_as_user
       get edit_user_registration_path